WIP move to integration tests
