+++ test/integration/helpers/html-utils.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import BpmnVisualization from '../../../../src/component/BpmnVisualization';
+import BpmnVisualization from '../../../src/component/BpmnVisualization';
 
 export class HtmlElementLookup {
   constructor(private bpmnVisualization: BpmnVisualization) {}
